+++ src/cli/steadybit-def-repo.ts
@@ -19,7 +19,9 @@ program
 
 program
   .command('check')
-  .description('Checks that the tasks and policies checked into the GitHub repository are valid.')
+  .description('Checks that the tasks and policies checked into the GitHub repository are valid. Assumes that the tasks and policies are checked in.')
+  .option('-v, --version <version>', 'The version to check. Must resolve to a git ref existing in the repository.')
+  .option('-d, --directory <directory>', 'The directory to search for task and policy files.', process.cwd())
   .action(check);
 
 program.parseAsync(process.argv);